+++ src/Udred/GUI/CaseFinder.fxml
@@ -18,7 +18,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<GridPane maxWidth="1.7976931348623157E308" styleClass="admin" stylesheets="@resources/css/MainStyle.css" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.CaseFinderController">
+<GridPane maxWidth="1.7976931348623157E308" styleClass="admin" stylesheets="@resources/css/MainStyle.css" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.GUI.CaseFinderController">
    <columnConstraints>
       <ColumnConstraints hgrow="SOMETIMES" maxWidth="10.0" minWidth="10.0" prefWidth="10.0" />
       <ColumnConstraints />
