+++ src/Udred/GUI/LoginWindow.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<GridPane hgap="5.0" vgap="5.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.LoginWindowController">
+<GridPane hgap="5.0" vgap="5.0" xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.GUI.LoginWindowController">
    <columnConstraints>
       <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="50.0" />
       <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="25.0" />
@@ -46,17 +46,17 @@
       </Label>
       <ImageView fitHeight="25.0" fitWidth="24.0" pickOnBounds="true" preserveRatio="true" GridPane.columnIndex="1" GridPane.rowIndex="4">
          <image>
-            <Image url="@resources/images/icons8-user-24.png" />
+            <Image url="@../resources/images/icons8-user-24.png" />
          </image>
       </ImageView>
       <ImageView fitHeight="25.0" fitWidth="19.0" pickOnBounds="true" preserveRatio="true" GridPane.columnIndex="1" GridPane.rowIndex="5">
          <image>
-            <Image url="@resources/images/icons8-lock-26.png" />
+            <Image url="@../resources/images/icons8-lock-26.png" />
          </image>
       </ImageView>
       <ImageView fitHeight="150.0" fitWidth="50.0" pickOnBounds="true" preserveRatio="true" GridPane.columnIndex="1" GridPane.columnSpan="2" GridPane.halignment="CENTER" GridPane.rowIndex="1" GridPane.valignment="CENTER">
          <image>
-            <Image url="@resources/images/EG.png" />
+            <Image url="@../resources/images/EG.png" />
          </image>
       </ImageView>
       <Button fx:id="btnLoginAdmin" mnemonicParsing="false" style="-fx-background-color: transparent;" GridPane.columnIndex="3" GridPane.rowIndex="8">
@@ -66,7 +66,7 @@
          <graphic>
             <ImageView fitHeight="40.0" fitWidth="40.0" pickOnBounds="true" preserveRatio="true" translateX="5.0" translateY="-5.0">
                <image>
-                  <Image url="@resources/images/icons8-rainbow-96.png" />
+                  <Image url="@../resources/images/icons8-rainbow-96.png" />
                </image>
                <viewport>
                   <Rectangle2D />
@@ -74,6 +74,6 @@
             </ImageView>
          </graphic>
       </Button>
-      <CheckBox fx:id="checkBoxRememberLogin" mnemonicParsing="false" text="Husk log ind" GridPane.columnIndex="2" GridPane.rowIndex="6" />
+      <CheckBox fx:id="checkBoxRememberLogin" mnemonicParsing="false" text="Husk Brugernavn" GridPane.columnIndex="2" GridPane.rowIndex="6" />
    </children>
 </GridPane>
