+++ src/Udred/GUI/PreLoader.fxml
@@ -4,11 +4,11 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.PreLoaderController">
+<AnchorPane xmlns="http://javafx.com/javafx/9.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Udred.GUI.PreLoaderController">
    <children>
       <ImageView fx:id="imageViewUni" fitHeight="400.0" fitWidth="600.0" pickOnBounds="true" preserveRatio="true">
          <image>
-            <Image url="@resources/images/Main.gif" />
+            <Image url="@../resources/images/Main.gif" />
          </image>
       </ImageView>
    </children>
